Correções ortográficas e melhorias de layout

+++ usuarios_manutencao.php
@@ -10,6 +10,7 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <link rel="stylesheet" href="assets/css/main.css" />
+    <link rel="stylesheet" href="assets/css/popup.css" />
 </head>
 <body>
   <?php
@@ -67,21 +68,34 @@ require "portal/menu.php";
 ?>
           <!-- Form -->
             <section>
-              <h3>Cadastrar novo usuário</h3>
+              <?php
+              //testa se é um cadastro de novo usuário, ou manutenção de existente
+              if ($login != ""){
+                  echo("<h3>Alterar dados do usuário</h3>");
+              } else {
+                  echo("<h3>Cadastrar novo usuário</h3>");
+              }
+              ?>
               <form method="post" accept-charset="utf-8">
                 <div class="row uniform 50%">
                   <div class="6u 12u$(xsmall)">
-                    <input type="text" name="nome" id="nome" value="<?php echo $nome; ?>" placeholder="Nome completo" />
+                    Login:  <input type="text" name="login" id="login" value="<?php echo $login; ?>" placeholder="Nome de usuário (login)" />
                   </div>
                   <div class="6u$ 12u$(xsmall)">
-                    <input type="text" name="login" id="login" value="<?php echo $login; ?>" placeholder="Nome de usuário (login)" />
+                    Nome do usuário: <input type="text" name="nome" id="nome" value="<?php echo $nome; ?>" placeholder="Nome completo" />
                   </div>
-                  <div class="6u 12u$(xsmall)">
-                    <input type="password" name="senha" id="senha" value="" placeholder="Senha" />
-                  </div>
-                  <div class="6u$ 12u$(xsmall)">
-                    <input type="password" name="senha-r" id="senha-r" value="" placeholder="Repita a senha" />
-                  </div>
+                  <?php
+                  //testa se é um cadastro de novo usuário
+                  // só deixa informar senha se for um novo cadastro
+                  if ($login == ""){
+                      echo("<div class=\"6u 12u$(xsmall)\">
+                      Senha:  <input type=\"password\" name=\"senha\" id=\"senha\" value=\"\" placeholder=\"Senha\" />
+                      </div>
+                      <div class=\"6u$ 12u$(xsmall)\">
+                       Repetir a senha:  <input type=\"password\" name=\"senha-r\" id=\"senha-r\" value=\"\" placeholder=\"Repita a senha\" />
+                      </div>");
+                  }
+                  ?>
                   <div class="4u 12u$(xsmall)">
                     <input type="radio" id="ind_Secretaria" value="ind_Secretaria" name="acesso" <?php echo $ind_Secretaria == 'S' ? 'checked' : false; ?>>
                     <label for="ind_Secretaria">Secretaria</label>
@@ -95,7 +109,7 @@ require "portal/menu.php";
                     <label for="ind_Professor">Professor</label>
                   </div>
                   <div class="6u$ 12u$(xsmall)">
-                    <input type="text" name="ra" id="ra" value="<?php echo $RA; ?>" placeholder="RA" />
+                    RA: <input type="text" name="ra" id="ra" value="<?php echo $RA; ?>" placeholder="RA" />
                   </div>
                   <div class="12u$">
                     <ul class="actions">
@@ -128,33 +142,35 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
     header('location:usuarios.php');
   }
 
-
-
   $login = isset($_POST["login"]) ? addslashes(trim($_POST["login"])) : FALSE;
   $nome = isset($_POST["nome"]) ? addslashes(trim($_POST["nome"])) : FALSE;
   $senha = isset($_POST["senha"]) ? md5(trim($_POST["senha"])) : FALSE;
+  $senha_r = isset($_POST["senha-r"]) ? md5(trim($_POST["senha-r"])) : FALSE;
   $ind_Secretaria = isset($_POST["acesso"]) ? ($_POST["acesso"] == 'ind_Secretaria' ? 'S' : 'N') : FALSE;
   $ind_Aluno = isset($_POST["acesso"]) ? ($_POST["acesso"] == 'ind_Aluno' ? 'S' : 'N') : FALSE;
   $ind_Professor = isset($_POST["acesso"]) ? ($_POST["acesso"] == 'ind_Professor' ? 'S' : 'N') : FALSE;
   $RA = isset($_POST["ra"]) ? addslashes(trim($_POST["ra"])) : FALSE;
 
-
-  if (isset($_POST['Incluir'])) {
-    mysqli_query($conn, "INSERT INTO usuarios (login, nome, senha, ind_Secretaria, ind_Aluno, ind_Professor, RA)" .
+  if ($senha != $senha_r){
+    echo("Senhas não coincidem!");
+  } else {
+    if (isset($_POST['Incluir'])) {
+      mysqli_query($conn, "INSERT INTO usuarios (login, nome, senha, ind_Secretaria, ind_Aluno, ind_Professor, RA)" .
                 " VALUES ('". $login . "', '" . $nome . "', '" . $senha . "', '" . $ind_Secretaria . "', '" . $ind_Aluno . "', '" . $ind_Professor . "', '" . $RA . "') ");
-  } elseif (isset($_POST['Alterar'])) {
-    mysqli_query($conn, "update usuarios set nome  = '" .  $nome . "'," .
-                "                    senha  = '" .  $senha . "'," .
-                "                    ind_Secretaria  = '" .  $ind_Secretaria . "'," .
-                "                    ind_Aluno  = '" .  $ind_Aluno . "'," .
-                "                    ind_Professor  = '" .  $ind_Professor . "'," .
-                "                    ra  = '" .  $RA . "' " .
-                " where login = '". $login . "'");
-  } elseif (isset($_POST['Excluir'])) {
-    mysqli_query($conn, "delete from usuarios where login = '". $login . "'");
+    } elseif (isset($_POST['Alterar'])) {
+      mysqli_query($conn, "update usuarios set nome  = '" .  $nome . "'," .
+                  "                    senha  = '" .  $senha . "'," .
+                  "                    ind_Secretaria  = '" .  $ind_Secretaria . "'," .
+                  "                    ind_Aluno  = '" .  $ind_Aluno . "'," .
+                  "                    ind_Professor  = '" .  $ind_Professor . "'," .
+                  "                    ra  = '" .  $RA . "' " .
+                  " where login = '". $login . "'");
+    } elseif (isset($_POST['Excluir'])) {
+      mysqli_query($conn, "delete from usuarios where login = '". $login . "'");
+    }
+    header('location:usuarios.php');
   }
 
-  header('location:usuarios.php');
 }
 ?>
